requirements.txt
# IF.witness CLI Requirements
# Python 3.10+

# CLI framework
click>=8.1.0

# HTTP client for vMix API
requests>=2.31.0

# OBS WebSocket client
obs-websocket-py>=1.0

# YAML config for vMix and OBS instances
PyYAML>=6.0.0

# Cryptography (Ed25519 signatures)
cryptography>=41.0.0

# PDF report generation (for compliance exports)
reportlab>=4.0.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0
